GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Virtual branch (refs/gitbutler/Virtual-branch)
   - pkg/hashmap/map.go
   - pkg/set/set_test.go
 - Change the type of Map.size and fixed the Map size 0 bug

The type of Map.size changes to uint32 for better memory use

Fixes bug in calculating index for Map with size 0
 (refs/gitbutler/Change-the-type-of-Map-size-and-fixed-the-Map-size-0-bug-The-type-of-Map-size-changes-to-uint32-for-better-memory-use-Fixes-bug-in-calculating-index-for-Map-with-size-0-)
   branch head: 4e06899513c3ad197f5bfba5d6ef988d23ce2da1

Your previous branch was: refs/heads/main

The sha for that commit was: ee15a5dd7785518689a6eedfbbb4c66eb255c57c

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch